kube/second/flask-react-sql/backend/requirements.txt
Flask
Flask-CORS
mysql-connector-python